Reworked the solution to leverage docker swarm encryption
Since docker swarms basically creates a rolling key VPN with it's overlay
network I decided to leverage that by running the tor hidden service and kali
container on one node and a simple gateway on another. This way if/when the
gateway gets taken down for sending probes it's easy to setup a new one. Only
problem is, cap_add which is needed to change kali service gateway is not
currently implemented so this feature is on hold (see
https://github.com/moby/moby/issues/25885) for more details. While it's worked
out one could still use a VPN bridge between the main server and it's own
jumpbox.